+++ aws-sdk-core/lib/aws-sdk-core/databasemigrationservice.rb
@@ -1,0 +1,5 @@
+Aws.add_service(:DatabaseMigrationService, {
+  api: "#{Aws::API_DIR}/dms/2016-01-01/api-2.json",
+  docs: "#{Aws::API_DIR}/dms/2016-01-01/docs-2.json",
+  examples: "#{Aws::API_DIR}/dms/2016-01-01/examples-1.json",
+})